requirements.txt
streamlit
pandas
numpy
scikit-learn
joblib
tensorflow
Pillow
imbalanced-learn